+++ mRemoteNG/UI/Window/ConnectionWindow.cs
@@ -77,7 +77,6 @@ namespace mRemoteNG.UI.Window
             cmenTabFullscreen.Click += (sender, args) => ToggleFullscreen();
             cmenTabSmartSize.Click += (sender, args) => ToggleSmartSize();
             cmenTabViewOnly.Click += (sender, args) => ToggleViewOnly();
-            cmenTabScreenshot.Click += (sender, args) => CreateScreenshot();
             cmenTabStartChat.Click += (sender, args) => StartChat();
             cmenTabTransferFile.Click += (sender, args) => TransferFile();
             cmenTabRefreshScreen.Click += (sender, args) => RefreshScreen();
@@ -268,7 +267,6 @@ namespace mRemoteNG.UI.Window
             cmenTabFullscreen.Text = Language.Fullscreen;
             cmenTabSmartSize.Text = Language.SmartSize;
             cmenTabViewOnly.Text = Language.ViewOnly;
-            cmenTabScreenshot.Text = Language.Screenshot;
             cmenTabStartChat.Text = Language.StartChat;
             cmenTabTransferFile.Text = Language.TransferFile;
             cmenTabRefreshScreen.Text = Language.RefreshScreen;
@@ -779,15 +777,6 @@ namespace mRemoteNG.UI.Window
             {
                 Runtime.MessageCollector.AddExceptionMessage("RenameTab (UI.Window.ConnectionWindow) failed", ex);
             }
-        }
-
-        private void CreateScreenshot()
-        {
-            cmenTab.Close();
-            Application.DoEvents();
-            //var selectedTab = (ConnectionTab)GetInterfaceControl()?.Parent;
-            if (TabHelper.Instance.CurrentTab == null) return;
-            Windows.ScreenshotForm.AddScreenshot(MiscTools.TakeScreenshot(TabHelper.Instance.CurrentTab));
         }
 
         #endregion